+++ test/set_state_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:set_state/set_state.dart';
 
-import '../example/lib/main.dart';
+import '../example/lib/src/my_app.dart';
 
 void main() {
   test('adds one to input values', () {